+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to
   - [#3203](https://github.com/bpftrace/bpftrace/pull/3203)
 - Remove length limitations for the `buf` builtin function
   - [#3249](https://github.com/bpftrace/bpftrace/pull/3249)
+- Faster map access for keyless maps by using BPF_MAP_TYPE_ARRAY
+  - [#3300](https://github.com/bpftrace/bpftrace/pull/3300)
 #### Deprecated
 #### Removed
 - Remove the `-dd` CLI option